13: Start doing project structure changes.
