+++ NugetAnalyzer.Web.Infrastructure/Models/GitHubUserClaimTypes.cs
@@ -1,0 +1,13 @@
+﻿namespace NugetAnalyzer.Web.Infrastructure.Models
+{
+    /// <summary>
+    /// Defines constants for GitHub user information keys.
+    /// </summary>
+    public class GitHubUserClaimTypes
+    {
+        public const string Id = "id";
+        public const string UserName = "login";
+        public const string Url = "html_url";
+        public const string AvatarUrl = "avatar_url";
+    }
+}
