+++ NugetAnalyzer.Web.Infrastructure/Models/OAuthSourceNames.cs
@@ -1,0 +1,7 @@
+﻿namespace NugetAnalyzer.Web.Models
+{
+    public static class OAuthSourceNames
+    {
+        public const string GitHubSourceName = "GitHub";
+    }
+}
