+++ NugetAnalyzer.Web.Infrastructure/Options/GitHubEndPointsOptions.cs
