+++ NugetAnalyzer.Web.Infrastructure/Options/GitHubSecretsOptions.cs
