+++ NugetAnalyzer.Web.Infrastructure2/Options/GitHubEndPointsOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace NugetAnalyzer.Web.Options
+{
+    public class GitHubEndPointsOptions
+    {
+        public string AuthorizationEndpoint { get; set; }
+
+        public string TokenEndpoint { get; set; }
+
+        public string UserInformationEndpoint { get; set; }
+    }
+}
