+++ NugetAnalyzer.Web/NugetAnalyzer.Web.csproj
@@ -23,6 +23,7 @@
     <ProjectReference Include="..\NugetAnalyzer.BLL\NugetAnalyzer.BLL.csproj" />
     <ProjectReference Include="..\NugetAnalyzer.Common\NugetAnalyzer.Common.csproj" />
     <ProjectReference Include="..\NugetAnalyzer.Dtos\NugetAnalyzer.Dtos.csproj" />
+    <ProjectReference Include="..\NugetAnalyzer.Web.Infrastructure\NugetAnalyzer.Web.Infrastructure.csproj" />
   </ItemGroup>
 
   <ItemGroup>
