--- NugetAnalyzer.Web/ServiceCollectionExtensions/ConverterExtensions.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using NugetAnalyzer.Dtos.Converters;
-
-namespace NugetAnalyzer.Web.ServiceCollectionExtensions
-{
-    public static class ConverterExtensions
-    {
-        public static void AddConverters(this IServiceCollection services)
-        {
-            services.AddScoped<ProfileConverter>();
-            services.AddScoped<UserConverter>();
-        }
-    }
-}
